Adicionando Permissoes através do Token

+++ DigitalBank.Api/Controllers/PersonController.cs
@@ -1,24 +1,33 @@
 ﻿using DigitalBank.Application.Contracts.Services;
 using DigitalBank.Application.DTOs.Person;
+using DigitalBank.Domain.Contracts.Authentication;
 using DigitalBank.Domain.FiltersDb;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DigitalBank.Api.Controllers;
 
 [Route("api/[controller]")]
-[ApiController]
-public class PersonController : ControllerBase
+public class PersonController : BaseController
 {
     private readonly IPersonService _personService;
+    private readonly ICurrentUser _currentUser;
+    private readonly List<string> _permissionUser;
+    private List<string> _permissionNeeded = new List<string>() { "Admin" };
 
-    public PersonController(IPersonService personService)
+    public PersonController(IPersonService personService, ICurrentUser currentUser)
     {
         _personService = personService;
+        _currentUser = currentUser;
+        _permissionUser = _currentUser.Permissions.Split(",").ToList() ?? new List<string>();
     }
 
     [HttpPost]
     public async Task<IActionResult> CreateAsync([FromBody] CreatePersonDto personDto)
     {
+        _permissionNeeded.Add("User");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.CreateAsync(personDto);
         if (result.Success)
             return CreatedAtAction(nameof(GetByIdAsync), new {id = result.Data.Id}, result.Data);
@@ -28,6 +37,10 @@ public class PersonController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteAsync(int id)
     {
+        _permissionNeeded.Add("User");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.DeleteAsync(id);
         if(result.Success)
             return NoContent();
@@ -37,6 +50,9 @@ public class PersonController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAllAsync()
     {
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.GetAllAsync();
         if (result.Success)
             return Ok(result.Data);
@@ -46,6 +62,10 @@ public class PersonController : ControllerBase
     [ActionName(nameof(GetByIdAsync))]
     public async Task<IActionResult> GetByIdAsync(int id)
     {
+        _permissionNeeded.Add("User");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.GetByIdAsync(id);
         if (result.Success)
             return Ok(result.Data);
@@ -55,6 +75,10 @@ public class PersonController : ControllerBase
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateAsync(int id, [FromBody] UpdatePersonDto personDto)
     {
+        _permissionNeeded.Add("User");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.UpdateAsync(id, personDto);
         if(result.Success)
             return NoContent();
@@ -66,6 +90,9 @@ public class PersonController : ControllerBase
     [HttpGet("paged")]
     public async Task<IActionResult> GetPagedAsync([FromQuery] PersonFilterDb personFilterDb)
     {
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+
         var result = await _personService.GetPagedAsync(personFilterDb);
         if(result.Success)
             return Ok(result.Data);
